requirements.txt
flask==3.0.3
firebase-admin==6.5.0
psycopg2-binary==2.9.9
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
